Add - Session and styles

+++ src/public/js/cart.js
@@ -1,6 +1,6 @@
 const carritoId = localStorage.getItem('cart-id');
 const addToCart = document.querySelectorAll('.addToCart');
-const deleteToCart = document.querySelectorAll('.deleteToCart');
+const deleteFromCart = document.querySelectorAll('.deleteFromCart');
 const deleteCart = document.querySelectorAll('.deleteCart');
 const cartIcon = document.getElementById('cartIcon');
 
@@ -37,29 +37,85 @@ addToCart.forEach(button => {
     const productId = button.getAttribute('data-productid');
 
     addProductToCart(productId);
+
+    Swal.fire({
+      title: 'Producto añadido al carrito',
+      text: 'El producto ha sido añadido al carrito de compras.',
+      icon: 'success',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Ir al carrito',
+      cancelButtonText: 'Seguir comprando',
+    }).then(result => {
+      if (result.isConfirmed) {
+        window.location.href = `/carts/${carritoId}`;
+      }
+    });
   });
 });
 
-deleteToCart.forEach(button => {
+deleteFromCart.forEach(button => {
   button.addEventListener('click', () => {
     const productId = button.getAttribute('data-productid');
-    console.log(productId);
-    deleteProductToCart(productId);
+    Swal.fire({
+      title: 'Eliminar del carrito',
+      text: '¿Estás seguro de que quieres eliminar este producto del carrito?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then(result => {
+      if (result.isConfirmed) {
+        deleteProductToCart(productId);
+        Swal.fire({
+          title: 'Carrito eliminado',
+          text: 'El producto ha sido eliminado exitosamente.',
+          icon: 'success',
+          confirmButtonColor: '#3085d6',
+          confirmButtonText: 'Aceptar',
+        }).then(() => {
+          window.location.href = `/carts/${carritoId}`;
+        });
+      }
+    });
   });
 });
 
 deleteCart.forEach(button => {
   button.addEventListener('click', () => {
     const cartId = button.getAttribute('data-cartid');
-    console.log(cartId);
-    deleteAllCart(cartId);
+    Swal.fire({
+      title: 'Eliminar carrito',
+      text: '¿Estás seguro de que quieres eliminar este carrito?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      confirmButtonText: 'Eliminar',
+      cancelButtonText: 'Cancelar',
+    }).then(result => {
+      if (result.isConfirmed) {
+        deleteAllCart(cartId);
+        Swal.fire({
+          title: 'Carrito eliminado',
+          text: 'El carrito ha sido eliminado exitosamente.',
+          icon: 'success',
+          confirmButtonColor: '#3085d6',
+          confirmButtonText: 'Aceptar',
+        }).then(() => {
+          window.location.href = `/carts/${carritoId}`;
+        });
+      }
+    });
   });
 });
 
 function addProductToCart(productId) {
   const products = { products: { product: productId } };
   const cartId = localStorage.getItem('cart-id');
-  console.log(cartId);
   fetch(`/carts/${cartId}/product/${productId}`, {
     method: 'PUT',
     headers: {
